reusing actions via request methods
