+++ 21---Building-a-Multi-Page-SPA-with-React-Router/new_project_setup/12-adv-starting-project/frontend/src/App.js
@@ -2,13 +2,14 @@ import { RouterProvider, createBrowserRouter} from "react-router-dom";
 import HomePage from "./pages/Home";
 import EventsPage from "./pages/Events";
 import EventDetailPage, {loader as eventDetailLoader } from "./pages/EventDetail";
-import NewEventPage, { action as newEventAction} from "./pages/NewEvent";
+import NewEventPage from "./pages/NewEvent";
 import EditEventPage from "./pages/EditEvent";
 import EventsRootLayout from "./pages/EventsRoot";
 import RootLayout from "./pages/Root";
 import { loader as eventsLoader } from "./pages/Events";
 import { action as deleteEventAction } from "./pages/EventDetail";
 import ErrorPage from "./pages/Error";
+import { action as manipulateEventAction } from "./components/EventForm";
 
 const router = createBrowserRouter([
   {
@@ -41,14 +42,15 @@ const router = createBrowserRouter([
               },
               {
                 path: "edit",
-                element: <EditEventPage />
+                element: <EditEventPage />,
+                action: manipulateEventAction
               }    
             ]
           },
           { 
             path: "new",
             element: <NewEventPage />,
-            action: newEventAction
+            action: manipulateEventAction
           }
         ]
       },
